requirements.txt
fastapi
uvicorn
watchdog
pdfplumber
python-docx
pytesseract
pillow
langchain
faiss-cpu
streamlit
pydantic
PyYAML
requests
Jinja2
reportlab
sentence-transformers
scikit-learn
python-multipart
langdetect
langgraph